requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pillow>=10.0.0
python-dotenv>=1.0.0
uuid>=1.30
dataclasses-json>=0.6.0
typing-extensions>=4.8.0